refactor: 중간 폴더 fakeapp 제거
