+++ src/pages/Simulation/PhishingSimulations/LoanInvestment/FakeAllowPermissionPage/components/PermissionItem.tsx
