+++ src/pages/Simulation/PhishingSimulations/LoanInvestment/FakeAllowPermissionPage/components/PermissionSwitch.tsx
@@ -1,4 +1,4 @@
-import { Switch } from "../../../../../../../components/ui/switch";
+import { Switch } from "../../../../../../components/ui/switch";
 
 interface PermissionSwitchProps{
     isChecked:boolean;
