+++ src/pages/Simulation/PhishingSimulations/LoanInvestment/FakeHomePage/components/FakeLoanItem.tsx
