+++ src/pages/Simulation/PhishingSimulations/LoanInvestment/FakeLandingPage/FakeAppLandingPage.tsx
