+++ src/pages/Simulation/PhishingSimulations/LoanInvestment/FakeLoanInvestPage/components/LoanInfoInput.tsx
